+++ waspc/data/Generator/templates/server/src/auth/providers/local/signup.ts
@@ -5,7 +5,10 @@ import { createUser } from '../../utils.js'
 export default handleRejection(async (req, res) => {
   const userFields = req.body || {}
 
-  await createUser(userFields)
+  await createUser({
+    username: userFields.username,
+    password: userFields.password,
+  })
 
   return res.json({ success: true })
 })