+++ waspc/headless-test/examples/todoApp/src/client/App.tsx
@@ -1,6 +1,6 @@
 import { Link } from 'react-router-dom'
 
-import logout from '@wasp/auth/logout.js'
+import logout from '@wasp/auth/logout'
 import useAuth from '@wasp/auth/useAuth'
 import { useQuery } from '@wasp/queries'
 import getDate from '@wasp/queries/getDate'
@@ -17,9 +17,7 @@ export function App({ children }: any) {
         <h1 className="font-bold text-3xl mb-5">
           <Link to="/">ToDo App</Link>
         </h1>
-        <h2>
-          Your site was loaded at: {date?.toLocaleString()}
-        </h2>
+        <h2>Your site was loaded at: {date?.toLocaleString()}</h2>
         {user && (
           <div className="flex gap-3 items-center">
             <div>
@@ -34,9 +32,7 @@ export function App({ children }: any) {
         )}
       </header>
       <main>{children}</main>
-      <footer className="mt-8 text-center">
-        Created with Wasp
-      </footer>
+      <footer className="mt-8 text-center">Created with Wasp</footer>
     </div>
   )
 }