+++ waspc/packages/deploy/src/providers/fly/cmd/cmd.ts
@@ -1,14 +1,17 @@
 import { exit } from 'process';
 import { $, cd } from 'zx';
 import { CommonOps, getCommonOps } from '../helpers/CommonOps.js';
-import { buildDirExists, getCommandHelp, waspSays } from '../helpers/helpers.js';
+import { buildDirExists, waspSays } from '../helpers/helpers.js';
 import { deleteLocalToml, getTomlFilePaths, localTomlExists } from '../helpers/tomlFileHelpers.js';
-import { executeFlyCommand } from '../index.js';
 import { CmdOptions } from './CmdOptions.js';
 
 // Runs a command by copying down the project toml files, executing it, and copying it back up (just in case).
 // If the toml file does not exist, some commands will not run with additional args (e.g. -a <appname>).
 export async function cmd(flyctlArgs: string[], options: CmdOptions): Promise<void> {
+	if (options.org) {
+		flyctlArgs.push('--org', options.org);
+	}
+
 	waspSays(`Running ${options.context} command: flyctl ${flyctlArgs.join(' ')}`);
 
 	if (!buildDirExists(options.waspProjectDir)) {
@@ -23,7 +26,7 @@ export async function cmd(flyctlArgs: string[], options: CmdOptions): Promise<vo
 	await runFlyctlCommand(commonOps, flyctlArgs);
 }
 
-async function runFlyctlCommand(commonOps: CommonOps, flyctlArgs: string[]) {
+async function runFlyctlCommand(commonOps: CommonOps, flyctlArgs: string[]): Promise<void> {
 	commonOps.cdToBuildDir();
 	deleteLocalToml();
 	if (commonOps.tomlExistsInProject()) {
@@ -33,8 +36,6 @@ async function runFlyctlCommand(commonOps: CommonOps, flyctlArgs: string[]) {
 	try {
 		await $`flyctl ${flyctlArgs}`;
 	} catch {
-		waspSays('Error running command. Note: many commands require a toml file or a -a option specifying the app name.');
-		waspSays(`If you already have an app, consider running "${getCommandHelp(executeFlyCommand).replace('<cmd...>', 'config save -- -a <app-name>')}".`);
 		exit(1);
 	}