+++ waspc/src/Wasp/AppSpec/Valid.hs
@@ -6,12 +6,13 @@ module Wasp.AppSpec.Valid
     getApp,
     isAuthEnabled,
     doesUserEntityContainField,
+    getIdFieldFromCrudEntity,
   )
 where
 
 import Control.Monad (unless)
 import Data.List (find, group, groupBy, intercalate, sort, sortBy)
-import Data.Maybe (isJust)
+import Data.Maybe (fromJust, isJust)
 import Text.Read (readMaybe)
 import Text.Regex.TDFA ((=~))
 import Wasp.AppSpec (AppSpec)
@@ -25,10 +26,14 @@ import qualified Wasp.AppSpec.App.Auth as Auth
 import qualified Wasp.AppSpec.App.Db as AS.Db
 import qualified Wasp.AppSpec.App.Wasp as Wasp
 import Wasp.AppSpec.Core.Decl (takeDecls)
+import qualified Wasp.AppSpec.Crud as AS.Crud
+import Wasp.AppSpec.Entity (isFieldUnique)
 import qualified Wasp.AppSpec.Entity as Entity
 import qualified Wasp.AppSpec.Entity.Field as Entity.Field
 import qualified Wasp.AppSpec.Page as Page
 import Wasp.AppSpec.Util (isPgBossJobExecutorUsed)
+import Wasp.Generator.Crud (crudDeclarationToOperationsList)
+import qualified Wasp.Psl.Ast.Model as PslModel
 import qualified Wasp.SemanticVersion as SV
 import qualified Wasp.Version as WV
 
@@ -54,7 +59,8 @@ validateAppSpec spec =
           validateExternalAuthEntityHasCorrectFieldsIfExternalAuthIsUsed spec,
           validateDbIsPostgresIfPgBossUsed spec,
           validateApiRoutesAreUnique spec,
-          validateApiNamespacePathsAreUnique spec
+          validateApiNamespacePathsAreUnique spec,
+          validateCrudOperations spec
         ]
 
 validateExactlyOneAppExists :: AppSpec -> Maybe ValidationError
@@ -74,7 +80,8 @@ validateWaspVersion :: String -> [ValidationError]
 validateWaspVersion specWaspVersionStr = eitherUnitToErrorList $ do
   specWaspVersionRange <- parseWaspVersionRange specWaspVersionStr
   unless (SV.isVersionInRange WV.waspVersion specWaspVersionRange) $
-    Left $ incompatibleVersionError WV.waspVersion specWaspVersionRange
+    Left $
+      incompatibleVersionError WV.waspVersion specWaspVersionRange
   where
     -- TODO: Use version range parser from SemanticVersion when it is fully implemented.
 
@@ -143,14 +150,37 @@ validateAuthUserEntityHasCorrectFieldsIfUsernameAndPasswordAuthIsUsed spec = cas
   Just auth ->
     if not $ Auth.isUsernameAndPasswordAuthEnabled auth
       then []
-      else
-        let userEntity = snd $ AS.resolveRef spec (Auth.userEntity auth)
-            userEntityFields = Entity.getFields userEntity
-         in concatMap
-              (validateEntityHasField "app.auth.userEntity" userEntityFields)
-              [ ("username", Entity.Field.FieldTypeScalar Entity.Field.String, "String"),
-                ("password", Entity.Field.FieldTypeScalar Entity.Field.String, "String")
-              ]
+      else validationErrors
+    where
+      validationErrors = concat [usernameValidationErrors, passwordValidationErrors]
+      usernameValidationErrors
+        | not $ null usernameTypeValidationErrors = usernameTypeValidationErrors
+        | otherwise = usernameAttributeValidationErrors
+      passwordValidationErrors =
+        validateEntityHasField
+          userEntityName
+          authUserEntityPath
+          userEntityFields
+          ("password", Entity.Field.FieldTypeScalar Entity.Field.String, "String")
+      usernameTypeValidationErrors =
+        validateEntityHasField
+          userEntityName
+          authUserEntityPath
+          userEntityFields
+          ("username", Entity.Field.FieldTypeScalar Entity.Field.String, "String")
+      usernameAttributeValidationErrors
+        | isFieldUnique "username" userEntity == Just True = []
+        | otherwise =
+            [ GenericValidationError $
+                "The field 'username' on entity '"
+                  ++ userEntityName
+                  ++ "' (referenced by "
+                  ++ authUserEntityPath
+                  ++ ") must be marked with the '@unique' attribute."
+            ]
+      userEntityFields = Entity.getFields userEntity
+      authUserEntityPath = "app.auth.userEntity"
+      (userEntityName, userEntity) = AS.resolveRef spec (Auth.userEntity auth)
 
 validateAuthUserEntityHasCorrectFieldsIfEmailAuthIsUsed :: AppSpec -> [ValidationError]
 validateAuthUserEntityHasCorrectFieldsIfEmailAuthIsUsed spec = case App.auth (snd $ getApp spec) of
@@ -159,10 +189,10 @@ validateAuthUserEntityHasCorrectFieldsIfEmailAuthIsUsed spec = case App.auth (sn
     if not $ Auth.isEmailAuthEnabled auth
       then []
       else
-        let userEntity = snd $ AS.resolveRef spec (Auth.userEntity auth)
+        let (userEntityName, userEntity) = AS.resolveRef spec (Auth.userEntity auth)
             userEntityFields = Entity.getFields userEntity
          in concatMap
-              (validateEntityHasField "app.auth.userEntity" userEntityFields)
+              (validateEntityHasField userEntityName "app.auth.userEntity" userEntityFields)
               [ ("email", Entity.Field.FieldTypeComposite (Entity.Field.Optional Entity.Field.String), "String"),
                 ("password", Entity.Field.FieldTypeComposite (Entity.Field.Optional Entity.Field.String), "String"),
                 ("isEmailVerified", Entity.Field.FieldTypeScalar Entity.Field.Boolean, "Boolean"),
@@ -197,7 +227,7 @@ validateExternalAuthEntityHasCorrectFieldsIfExternalAuthIsUsed spec = case App.a
               externalAuthEntityFields = Entity.getFields externalAuthEntity
               externalAuthEntityValidationErrors =
                 concatMap
-                  (validateEntityHasField "app.auth.externalAuthEntity" externalAuthEntityFields)
+                  (validateEntityHasField externalAuthEntityName "app.auth.externalAuthEntity" externalAuthEntityFields)
                   [ ("provider", Entity.Field.FieldTypeScalar Entity.Field.String, "String"),
                     ("providerId", Entity.Field.FieldTypeScalar Entity.Field.String, "String"),
                     ("user", Entity.Field.FieldTypeScalar (Entity.Field.UserType userEntityName), userEntityName),
@@ -205,20 +235,23 @@ validateExternalAuthEntityHasCorrectFieldsIfExternalAuthIsUsed spec = case App.a
                   ]
               userEntityValidationErrors =
                 concatMap
-                  (validateEntityHasField "app.auth.userEntity" userEntityFields)
-                  [ ("externalAuthAssociations", Entity.Field.FieldTypeComposite $ Entity.Field.List $ Entity.Field.UserType externalAuthEntityName, externalAuthEntityName ++ "[]")
+                  (validateEntityHasField userEntityName "app.auth.userEntity" userEntityFields)
+                  [ ( "externalAuthAssociations",
+                      Entity.Field.FieldTypeComposite $ Entity.Field.List $ Entity.Field.UserType externalAuthEntityName,
+                      externalAuthEntityName ++ "[]"
+                    )
                   ]
            in externalAuthEntityValidationErrors ++ userEntityValidationErrors
 
-validateEntityHasField :: String -> [Entity.Field.Field] -> (String, Entity.Field.FieldType, String) -> [ValidationError]
-validateEntityHasField entityName entityFields (fieldName, fieldType, fieldTypeName) =
-  let maybeField = find ((== fieldName) . Entity.Field.fieldName) entityFields
+validateEntityHasField :: String -> String -> [Entity.Field.Field] -> (String, Entity.Field.FieldType, String) -> [ValidationError]
+validateEntityHasField entityName authEntityPath entityFields (fieldName, fieldType, fieldTypeName) =
+  let maybeField = findFieldByName fieldName entityFields
    in case maybeField of
         Just providerField
           | Entity.Field.fieldType providerField == fieldType -> []
         _ ->
           [ GenericValidationError $
-              "Expected an Entity referenced by " ++ entityName ++ " to have field '" ++ fieldName ++ "' of type '" ++ fieldTypeName ++ "'."
+              "Entity '" ++ entityName ++ "' (referenced by " ++ authEntityPath ++ ") must have field '" ++ fieldName ++ "' of type '" ++ fieldTypeName ++ "'."
           ]
 
 validateApiRoutesAreUnique :: AppSpec -> [ValidationError]
@@ -250,6 +283,47 @@ validateApiNamespacePathsAreUnique spec =
     namespacePaths = AS.ApiNamespace.path . snd <$> AS.getApiNamespaces spec
     duplicatePaths = map head $ filter ((> 1) . length) (group . sort $ namespacePaths)
 
+validateCrudOperations :: AppSpec -> [ValidationError]
+validateCrudOperations spec =
+  concat
+    [ concatMap checkIfAtLeastOneOperationIsUsedForCrud cruds,
+      concatMap checkIfSimpleIdFieldIsDefinedForEntity cruds
+    ]
+  where
+    cruds = AS.getCruds spec
+
+    checkIfAtLeastOneOperationIsUsedForCrud :: (String, AS.Crud.Crud) -> [ValidationError]
+    checkIfAtLeastOneOperationIsUsedForCrud (crudName, crud) =
+      if not . null $ crudOperations
+        then []
+        else [GenericValidationError $ "CRUD \"" ++ crudName ++ "\" must have at least one operation defined."]
+      where
+        crudOperations = crudDeclarationToOperationsList crud
+
+    checkIfSimpleIdFieldIsDefinedForEntity :: (String, AS.Crud.Crud) -> [ValidationError]
+    checkIfSimpleIdFieldIsDefinedForEntity (crudName, crud) = case (maybeIdField, maybeIdBlockAttribute) of
+      (Just _, Nothing) -> []
+      (Nothing, Just _) ->
+        [ GenericValidationError $
+            "Entity '"
+              ++ entityName
+              ++ "' (referenced by CRUD declaration '"
+              ++ crudName
+              ++ "') must have an ID field (specified with the '@id' attribute) and not a composite ID (specified with the '@@id' attribute)."
+        ]
+      _missingIdFieldWithoutBlockIdAttributeDefined ->
+        [ GenericValidationError $
+            "Entity '"
+              ++ entityName
+              ++ "' (referenced by CRUD declaration '"
+              ++ crudName
+              ++ "') must have an ID field (specified with the '@id' attribute)."
+        ]
+      where
+        maybeIdField = Entity.getIdField entity
+        maybeIdBlockAttribute = Entity.getIdBlockAttribute entity
+        (entityName, entity) = AS.resolveRef spec (AS.Crud.entity crud)
+
 -- | This function assumes that @AppSpec@ it operates on was validated beforehand (with @validateAppSpec@ function).
 -- TODO: It would be great if we could ensure this at type level, but we decided that was too much work for now.
 --   Check https://github.com/wasp-lang/wasp/pull/455 for considerations on this and analysis of different approaches.
@@ -280,4 +354,14 @@ doesUserEntityContainField spec fieldName = do
   auth <- App.auth (snd $ getApp spec)
   let userEntity = snd $ AS.resolveRef spec (Auth.userEntity auth)
   let userEntityFields = Entity.getFields userEntity
-  Just $ any (\field -> Entity.Field.fieldName field == fieldName) userEntityFields
+  Just $ isJust $ findFieldByName fieldName userEntityFields
+
+findFieldByName :: String -> [Entity.Field.Field] -> Maybe Entity.Field.Field
+findFieldByName name = find ((== name) . Entity.Field.fieldName)
+
+-- | This function assumes that @AppSpec@ it operates on was validated beforehand (with @validateAppSpec@ function).
+-- We validated that entity field exists, so we can safely use fromJust here.
+getIdFieldFromCrudEntity :: AppSpec -> AS.Crud.Crud -> PslModel.Field
+getIdFieldFromCrudEntity spec crud = fromJust $ Entity.getIdField crudEntity
+  where
+    crudEntity = snd $ AS.resolveRef spec (AS.Crud.entity crud)