+++ waspc/src/Wasp/Generator/ServerGenerator.hs
@@ -32,7 +32,7 @@ import StrongPath
     relfile,
     (</>),
   )
-import Wasp.AppSpec (AppSpec, getApis)
+import Wasp.AppSpec (AppSpec)
 import qualified Wasp.AppSpec as AS
 import qualified Wasp.AppSpec.App as AS.App
 import qualified Wasp.AppSpec.App.Auth as AS.App.Auth
@@ -50,13 +50,13 @@ import Wasp.Generator.Common
 import Wasp.Generator.ExternalCodeGenerator (genExternalCodeDir)
 import Wasp.Generator.FileDraft (FileDraft, createTextFileDraft)
 import Wasp.Generator.Monad (Generator)
-import qualified Wasp.Generator.Node.Version as NodeVersion
 import qualified Wasp.Generator.NpmDependencies as N
 import Wasp.Generator.ServerGenerator.ApiRoutesG (genApis)
 import Wasp.Generator.ServerGenerator.Auth.OAuthAuthG (depsRequiredByPassport)
 import Wasp.Generator.ServerGenerator.AuthG (genAuth)
 import qualified Wasp.Generator.ServerGenerator.Common as C
 import Wasp.Generator.ServerGenerator.ConfigG (genConfigFile)
+import Wasp.Generator.ServerGenerator.CrudG (genCrud)
 import Wasp.Generator.ServerGenerator.Db.Seed (genDbSeed, getPackageJsonPrismaSeedField)
 import Wasp.Generator.ServerGenerator.EmailSenderG (depsRequiredByEmail, genEmailSender)
 import Wasp.Generator.ServerGenerator.ExternalCodeGenerator (extServerCodeGeneratorStrategy, extSharedCodeGeneratorStrategy)
@@ -64,6 +64,7 @@ import Wasp.Generator.ServerGenerator.JobGenerator (depsRequiredByJobs, genJobEx
 import Wasp.Generator.ServerGenerator.JsImport (extImportToImportJson, getAliasedJsImportStmtAndIdentifier)
 import Wasp.Generator.ServerGenerator.OperationsG (genOperations)
 import Wasp.Generator.ServerGenerator.OperationsRoutesG (genOperationsRoutes)
+import qualified Wasp.Node.Version as NodeVersion
 import Wasp.Project.Db (databaseUrlEnvVarName)
 import Wasp.SemanticVersion (major)
 import Wasp.Util (toLowerFirst, (<++>))
@@ -89,6 +90,7 @@ genServer spec =
     <++> genEnvValidationScript
     <++> genExportedTypesDir spec
     <++> genApis spec
+    <++> genCrud spec
   where
     genFileCopy = return . C.mkTmplFd
 
@@ -172,7 +174,7 @@ npmDepsForWasp spec =
             ("prisma", show prismaVersion),
             -- TODO: Allow users to choose whether they want to use TypeScript
             -- in their projects and install these dependencies accordingly.
-            ("typescript", "^4.8.4"),
+            ("typescript", "^5.1.0"),
             ("@types/express", "^4.17.13"),
             ("@types/express-serve-static-core", "^4.17.13"),
             ("@types/node", "^18.11.9"),
@@ -258,24 +260,30 @@ genRoutesDir :: AppSpec -> Generator [FileDraft]
 genRoutesDir spec =
   -- TODO(martin): We will probably want to extract "routes" path here same as we did with "src", to avoid hardcoding,
   -- but I did not bother with it yet since it is used only here for now.
-  return
-    [ C.mkTmplFdWithDstAndData
-        (C.asTmplFile [relfile|src/routes/index.js|])
-        (C.asServerFile [relfile|src/routes/index.js|])
-        ( Just $
-            object
-              [ "operationsRouteInRootRouter" .= (operationsRouteInRootRouter :: String),
-                "isAuthEnabled" .= (isAuthEnabled spec :: Bool),
-                "areThereAnyCustomApiRoutes" .= (not . null $ getApis spec)
-              ]
-        )
-    ]
+  sequence [genRoutesIndex spec]
+
+genRoutesIndex :: AppSpec -> Generator FileDraft
+genRoutesIndex spec =
+  return $
+    C.mkTmplFdWithDstAndData
+      (C.asTmplFile [relfile|src/routes/index.js|])
+      (C.asServerFile [relfile|src/routes/index.js|])
+      (Just tmplData)
+  where
+    tmplData =
+      object
+        [ "operationsRouteInRootRouter" .= (operationsRouteInRootRouter :: String),
+          "isAuthEnabled" .= (isAuthEnabled spec :: Bool),
+          "areThereAnyCustomApiRoutes" .= (not . null $ AS.getApis spec),
+          "areThereAnyCrudRoutes" .= (not . null $ AS.getCruds spec)
+        ]
 
 genTypesAndEntitiesDirs :: AppSpec -> Generator [FileDraft]
 genTypesAndEntitiesDirs spec =
   return
     [ entitiesIndexFileDraft,
       taggedEntitiesFileDraft,
+      serializationFileDraft,
       typesIndexFileDraft
     ]
   where
@@ -289,6 +297,9 @@ genTypesAndEntitiesDirs spec =
         [relfile|src/_types/taggedEntities.ts|]
         [relfile|src/_types/taggedEntities.ts|]
         (Just $ object ["entities" .= allEntities])
+    serializationFileDraft =
+      C.mkSrcTmplFd
+        [relfile|_types/serialization.ts|]
     typesIndexFileDraft =
       C.mkTmplFdWithDstAndData
         [relfile|src/_types/index.ts|]
@@ -379,9 +390,10 @@ genExportedTypesDir spec =
 
 genMiddleware :: AppSpec -> Generator [FileDraft]
 genMiddleware spec =
-  return
-    [ C.mkTmplFd [relfile|src/middleware/index.ts|],
-      C.mkTmplFdWithData [relfile|src/middleware/globalMiddleware.ts|] (Just tmplData)
+  sequence
+    [ return $ C.mkTmplFd [relfile|src/middleware/index.ts|],
+      return $ C.mkTmplFdWithData [relfile|src/middleware/globalMiddleware.ts|] (Just tmplData),
+      genOperationsMiddleware spec
     ]
   where
     tmplData =
@@ -399,3 +411,13 @@ genMiddleware spec =
               "importStatement" .= maybe "" fst maybeGlobalMidlewareConfigFnImports,
               "importAlias" .= globalMiddlewareConfigFnAlias
             ]
+
+genOperationsMiddleware :: AppSpec -> Generator FileDraft
+genOperationsMiddleware spec =
+  return $
+    C.mkTmplFdWithDstAndData
+      (C.asTmplFile [relfile|src/middleware/operations.ts|])
+      (C.asServerFile [relfile|src/middleware/operations.ts|])
+      (Just tmplData)
+  where
+    tmplData = object ["isAuthEnabled" .= (isAuthEnabled spec :: Bool)]