+++ waspc/src/Wasp/Generator/WebAppGenerator.hs
@@ -33,12 +33,13 @@ import Wasp.Generator.Common
   )
 import qualified Wasp.Generator.ConfigFile as G.CF
 import Wasp.Generator.ExternalCodeGenerator (genExternalCodeDir)
-import Wasp.Generator.FileDraft
+import Wasp.Generator.FileDraft (FileDraft, createTextFileDraft)
+import qualified Wasp.Generator.FileDraft as FD
 import Wasp.Generator.Monad (Generator)
-import qualified Wasp.Generator.Node.Version as NodeVersion
 import qualified Wasp.Generator.NpmDependencies as N
 import Wasp.Generator.WebAppGenerator.AuthG (genAuth)
 import qualified Wasp.Generator.WebAppGenerator.Common as C
+import Wasp.Generator.WebAppGenerator.CrudG (genCrud)
 import Wasp.Generator.WebAppGenerator.ExternalCodeGenerator
   ( extClientCodeGeneratorStrategy,
     extSharedCodeGeneratorStrategy,
@@ -46,11 +47,13 @@ import Wasp.Generator.WebAppGenerator.ExternalCodeGenerator
 import Wasp.Generator.WebAppGenerator.JsImport (extImportToImportJson)
 import Wasp.Generator.WebAppGenerator.OperationsGenerator (genOperations)
 import Wasp.Generator.WebAppGenerator.RouterGenerator (genRouter)
+import qualified Wasp.Node.Version as NodeVersion
 import qualified Wasp.SemanticVersion as SV
 import Wasp.Util ((<++>))
 
 genWebApp :: AppSpec -> Generator [FileDraft]
 genWebApp spec = do
+  extClientCodeFileDrafts <- genExternalCodeDir extClientCodeGeneratorStrategy (AS.externalClientFiles spec)
   sequence
     [ genFileCopy [relfile|README.md|],
       genFileCopy [relfile|tsconfig.json|],
@@ -65,13 +68,14 @@ genWebApp spec = do
       genGitignore,
       genIndexHtml spec
     ]
-    <++> genPublicDir spec
     <++> genSrcDir spec
-    <++> genExternalCodeDir extClientCodeGeneratorStrategy (AS.externalClientFiles spec)
+    <++> return extClientCodeFileDrafts
     <++> genExternalCodeDir extSharedCodeGeneratorStrategy (AS.externalSharedFiles spec)
+    <++> genPublicDir spec extClientCodeFileDrafts
     <++> genDotEnv spec
     <++> genUniversalDir
     <++> genEnvValidationScript
+    <++> genCrud spec
   where
     genFileCopy = return . C.mkTmplFd
 
@@ -118,10 +122,10 @@ npmDepsForWasp spec =
   N.NpmDepsForWasp
     { N.waspDependencies =
         AS.Dependency.fromList
-          [ ("axios", "^0.27.2"),
-            ("react", "^17.0.2"),
-            ("react-dom", "^17.0.2"),
-            ("@tanstack/react-query", "^4.13.0"),
+          [ ("axios", "^1.4.0"),
+            ("react", "^18.2.0"),
+            ("react-dom", "^18.2.0"),
+            ("@tanstack/react-query", "^4.29.0"),
             ("react-router-dom", "^5.3.3"),
             -- The web app only needs @prisma/client (we're using the server's
             -- CLI to generate what's necessary, check the description in
@@ -135,16 +139,16 @@ npmDepsForWasp spec =
         AS.Dependency.fromList
           [ -- TODO: Allow users to choose whether they want to use TypeScript
             -- in their projects and install these dependencies accordingly.
-            ("vite", "^4.1.0"),
-            ("typescript", "^4.9.3"),
-            ("@types/react", "^17.0.53"),
-            ("@types/react-dom", "^17.0.19"),
+            ("vite", "^4.3.9"),
+            ("typescript", "^5.1.0"),
+            ("@types/react", "^18.0.37"),
+            ("@types/react-dom", "^18.0.11"),
             ("@types/react-router-dom", "^5.3.3"),
             ("@vitejs/plugin-react-swc", "^3.0.0"),
             ("dotenv", "^16.0.3"),
             -- NOTE: Make sure to bump the version of the tsconfig
             -- when updating Vite or React versions
-            ("@tsconfig/vite-react", "^1.0.1")
+            ("@tsconfig/vite-react", "^2.0.0")
           ]
           ++ depsRequiredForTesting
     }
@@ -172,7 +176,7 @@ depsRequiredForTesting =
     [ ("vitest", "^0.29.3"),
       ("@vitest/ui", "^0.29.3"),
       ("jsdom", "^21.1.1"),
-      ("@testing-library/react", "^12.1.5"),
+      ("@testing-library/react", "^14.0.0"),
       ("@testing-library/jest-dom", "^5.16.5"),
       ("msw", "^1.1.0")
     ]
@@ -184,18 +188,25 @@ genGitignore =
       (C.asTmplFile [relfile|gitignore|])
       (C.asWebAppFile [relfile|.gitignore|])
 
-genPublicDir :: AppSpec -> Generator [FileDraft]
-genPublicDir spec = do
-  return
-    [ genFaviconFd,
-      genManifestFd
-    ]
+genPublicDir :: AppSpec -> [FileDraft] -> Generator [FileDraft]
+genPublicDir spec extCodeFileDrafts =
+  return $
+    ifUserDidntProvideFile genFaviconFd
+      ++ ifUserDidntProvideFile genManifestFd
   where
     genFaviconFd = C.mkTmplFd (C.asTmplFile [relfile|public/favicon.ico|])
-    genManifestFd =
-      let tmplData = object ["appName" .= (fst (getApp spec) :: String)]
-          tmplFile = C.asTmplFile [relfile|public/manifest.json|]
-       in C.mkTmplFdWithData tmplFile tmplData
+    genManifestFd = C.mkTmplFdWithData tmplFile tmplData
+      where
+        tmplData = object ["appName" .= (fst (getApp spec) :: String)]
+        tmplFile = C.asTmplFile [relfile|public/manifest.json|]
+
+    ifUserDidntProvideFile fileDraft =
+      if checkIfFileDraftExists fileDraft
+        then []
+        else [fileDraft]
+
+    checkIfFileDraftExists = (`elem` existingDstPaths) . FD.getDstPath
+    existingDstPaths = map FD.getDstPath extCodeFileDrafts
 
 genIndexHtml :: AppSpec -> Generator FileDraft
 genIndexHtml spec =
@@ -218,15 +229,15 @@ genIndexHtml spec =
 genSrcDir :: AppSpec -> Generator [FileDraft]
 genSrcDir spec =
   sequence
-    [ copyTmplFile [relfile|logo.png|],
-      copyTmplFile [relfile|config.js|],
-      copyTmplFile [relfile|queryClient.js|],
-      copyTmplFile [relfile|utils.js|],
-      copyTmplFile [relfile|types.ts|],
-      copyTmplFile [relfile|vite-env.d.ts|],
+    [ genFileCopy [relfile|logo.png|],
+      genFileCopy [relfile|config.js|],
+      genFileCopy [relfile|queryClient.js|],
+      genFileCopy [relfile|utils.js|],
+      genFileCopy [relfile|types.ts|],
+      genFileCopy [relfile|vite-env.d.ts|],
       -- Generates api.js file which contains token management and configured api (e.g. axios) instance.
-      copyTmplFile [relfile|api.ts|],
-      copyTmplFile [relfile|storage.ts|],
+      genFileCopy [relfile|api.ts|],
+      genFileCopy [relfile|storage.ts|],
       genRouter spec,
       genIndexJs spec
     ]
@@ -234,7 +245,7 @@ genSrcDir spec =
     <++> genEntitiesDir spec
     <++> genAuth spec
   where
-    copyTmplFile = return . C.mkSrcTmplFd
+    genFileCopy = return . C.mkSrcTmplFd
 
 genEntitiesDir :: AppSpec -> Generator [FileDraft]
 genEntitiesDir spec = return [entitiesIndexFileDraft]